+++ lib/sapience/formatters/raw.rb
@@ -4,7 +4,7 @@ module Sapience
     class Raw < Base
       # Returns log messages in Hash format
       def call(log, logger)
-        log.to_h(log_host ? logger.host : nil, log_application ? logger.application : nil)
+        log.to_h(log_host ? logger.host : nil, log_application ? logger.app_name : nil)
       end
     end
   end