+++ lib/sapience/log.rb
@@ -153,7 +153,7 @@ module Sapience
     end
 
     # Returns [Hash] representation of this log entry
-    def to_h(host = Sapience.config.host, application = Sapience.config.application) # rubocop:disable AbcSize, CyclomaticComplexity, PerceivedComplexity, LineLength
+    def to_h(host = Sapience.config.host, app_name = Sapience.app_name) # rubocop:disable AbcSize, CyclomaticComplexity, PerceivedComplexity, LineLength
       # Header
       h               = {
         name:        name,
@@ -163,9 +163,9 @@ module Sapience
         level:       level,
         level_index: level_index,
       }
-      h[:host]        = host if host
-      h[:application] = application if application
-      file, line      = file_name_and_line
+      h[:host]     = host if host
+      h[:app_name] = app_name if app_name
+      file, line   = file_name_and_line
       if file
         h[:file] = file
         h[:line] = line.to_i