+++ spec/lib/sapience/appender/datadog_spec.rb
@@ -15,6 +15,8 @@ describe Sapience::Appender::Datadog do
   let(:metric) { "my/own/metric" }
 
   before do
+    allow(Sapience).to receive(:environment).and_return("rspec")
+    Sapience.configure { |c| c.app_name = "datadog" }
     allow(::Datadog::Statsd).to receive(:new).and_return(statsd)
     allow(statsd).to receive(:batch).and_yield
   end
@@ -28,14 +30,18 @@ describe Sapience::Appender::Datadog do
       let(:url) { nil }
 
       it "sets the default url" do
-        expect(::Datadog::Statsd).to receive(:new).with("localhost", 8125, tags: nil)
+        expect(::Datadog::Statsd)
+          .to receive(:new)
+          .with("localhost", 8125, namespace: "datadog.rspec", tags: nil)
         subject.provider
       end
     end
 
     context "with url provided" do
       it "sets the url" do
-        expect(::Datadog::Statsd).to receive(:new).with("0.0.0.0", 2222, tags: nil)
+        expect(::Datadog::Statsd)
+          .to receive(:new)
+          .with("0.0.0.0", 2222, namespace: "datadog.rspec", tags: nil)
         subject.provider
       end
     end
@@ -43,7 +49,9 @@ describe Sapience::Appender::Datadog do
     context "with tags provided" do
       let(:tags) { "tag1:true" }
       it "sets the url" do
-        expect(::Datadog::Statsd).to receive(:new).with("0.0.0.0", 2222, tags: "tag1:true")
+        expect(::Datadog::Statsd)
+          .to receive(:new)
+          .with("0.0.0.0", 2222, namespace: "datadog.rspec", tags: tags)
         subject.provider
       end
     end